requirements.txt
# Core
python-dotenv==1.1.0
google-genai==1.14.0

# Technical Analysis
pandas==2.2.3
pandas-ta==0.3.14b1
numpy==2.2.4

# MetaTrader 5
MetaTrader5==5.0.4200

# MCP (Model Context Protocol)
mcp==1.8.0

# Web Dashboard
flask==3.1.0
flask-socketio==5.5.1

# Utilities
schedule==1.2.2
rich==13.9.4
requests==2.32.3
